+++ micropip/wheelinfo.py
@@ -180,7 +180,7 @@ class WheelInfo:
         return requires
 
     async def _fetch_bytes(self, url: str, fetch_kwargs: dict[str, Any]):
-        if self.parsed_url.scheme not in ("https", "http", "emfs"):
+        if self.parsed_url.scheme not in ("https", "http", "emfs", "file"):
             # Don't raise ValueError it gets swallowed
             raise TypeError(
                 f"Cannot download from a non-remote location: {url!r} ({self.parsed_url!r})"
@@ -196,7 +196,7 @@ class WheelInfo:
                 raise e
             else:
                 raise ValueError(
-                    f"Can't fetch wheel from '{url}'. "
+                    f"Can't fetch wheel from {url!r}. "
                     "One common reason for this is when the server blocks "
                     "Cross-Origin Resource Sharing (CORS). "
                     "Check if the server is sending the correct 'Access-Control-Allow-Origin' header."