+++ tests/test_freeze.py
@@ -64,3 +64,31 @@ async def test_freeze_fix_depends(
     assert pkg_metadata["imports"] == toplevel[0]
     assert dep1_metadata["imports"] == toplevel[1]
     assert dep2_metadata["imports"] == toplevel[2]
+
+
+def test_freeze_lockfile_compat(selenium_standalone_micropip, wheel_catalog, tmp_path):
+    from pyodide_lock import PyodideLockSpec
+
+    selenium = selenium_standalone_micropip
+    snowball_wheel = wheel_catalog.get("snowballstemmer")
+    url = snowball_wheel.url
+
+    lockfile_content = selenium.run_async(
+        f"""
+        await micropip.install("{url}")
+        micropip.freeze()
+    """
+    )
+
+    lockfile_path = tmp_path / "lockfile.json"
+    with open(lockfile_path, "w") as f:
+        f.write(lockfile_content)
+
+    lockfile = PyodideLockSpec.from_json(lockfile_path)
+    assert lockfile.packages["snowballstemmer"].file_name == url
+    assert lockfile.packages["snowballstemmer"].name == "snowballstemmer"
+    assert lockfile.packages["snowballstemmer"].depends == []
+    assert lockfile.packages["snowballstemmer"].imports == ["snowballstemmer"]
+    assert lockfile.packages["snowballstemmer"].install_dir == "site"
+    assert not lockfile.packages["snowballstemmer"].unvendored_tests
+    assert lockfile.packages["snowballstemmer"].version == snowball_wheel.version