+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.7.2] - 2024/11/26
+
+### Fixed
+
+ - Fix a bug that prevented some wasm32 wheel to be recognized as compatible with pyodide
+  [#159](https://github.com/pyodide/micropip/pull/159)
+
+## [0.7.1] - 2024/11/11
+
+### Fixed
+
+- Fixed bug that prevented package installation from `file:` path in Node.js environment.
+  [#155](https://github.com/pyodide/micropip/pull/155)
+
+## [0.7.0] - 2024/11/09
+
 ### Fixed
 
 - `micropip.install` now works with simple http indexes that use relative